Refs #124. Moved xml vulnerabilities testing to its own module
